+++ ker/include/ker/common_protocol.h
@@ -63,7 +63,7 @@ uint32_t            common_pack_req_describe(char* _buffer, uint16_t _size, uint
 
 uint32_t            common_pack_req_select(char* _buffer, uint16_t _size, uint16_t _remoteId, const char* _tableName, uint16_t _key);
 
-uint32_t            common_pack_req_insert(char* _buffer, uint16_t _size, uint16_t _remoteId, const char* _tableName, uint16_t _key, const char* _value, uint32_t _timestamp);
+uint32_t            common_pack_req_insert(char* _buffer, uint16_t _size, uint16_t _remoteId, const char* _tableName, uint16_t _key, const char* _value, uint64_t _timestamp);
 
 uint32_t            common_pack_res_create(char* _buffer, uint16_t _size, uint16_t _remoteId, const cx_err_t* _err);
